Computer player add!!!
For that AI module is imported
In game_cycle function
in while loop of choosing dices to re-roll and  adding points or striking figures
added if statement that checks who plays: Human or computer,
and executes that loop in old way for human
and new (with function from AI module) for computer.
Unfortunatelly computer is not able yet to choose dices to re-roll
so if the game has more than one throw, computer takes figures only from first throw.